Introduce crane for rust nix targets
This adds the [crane](https://crane.dev/introduction.html) library for
our nix builds of Rust targets.

I haven't converted everything to crane, but I did convert the
following CI checks:

- cargo clippy
- cargo audit
- cargo fmt

As well as both `schema-gen` and `spec-proxy` builds.

I'll catch the rest of the package builds in the near future, but I
wanted to move on to the next high-value targets first (migrations,
cargo tests, wasm builds, and javascript builds).

This also starts the process of (re-)splitting out the CI jobs into
smaller, more focused units: since we're relying on the remote nix
cache rather than the local sccache, we don't need to worry about
trying to push everything into a smaller number of CI jobs so that
they are better able to take advantage of the local cache.

Changelog: internal